Tested out methods for fixing rocket jumper whistle, none worked in a neat way. Temp patch currently applied on the servers is added to repo
Temp patch is just a nuclear fix for rocket jumper whistle getting stuck playing, it simply forces it to turn off for all players who have no crit and checks every second. Kind of a trash fix but should do the job for now (sound will only ever play for at most a second when it shouldn't).

Will update if a better solution is found.