Big update to tournament plugin in preparation for upcoming mgatf tournament
Added support for automatic sudden death as well as domination ends to duels.

This means players won't need to manually forfeit, and keep track of the timelimit when duelling in the tournament.

To enable this add "mga_TourneyPlug" to the "AutoRunPlugins" config element

And add "Tournament" : true to the config table

This is defined in the mga_config.nut file.